website/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0

# PyTorch CPU
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.2.0
torchvision>=0.17.0

# InsightFace pour le face swap
insightface>=0.7.3
onnxruntime>=1.16.0